Moved clair-c2py/ to clair/c2py/ & moved codegen/, ir/, and module_info.{cpp,hpp} into clair/
