+++ src/tools/clair/CMakeLists.txt
@@ -1,0 +1,10 @@
+add_tool(clair-c2py DIR c2py
+  "codegen/*.cpp"
+  "ir/*.cpp"
+  "module_info.cpp"
+)
+# add_tool(clair-f2py DIR f2py
+#   "codegen/*.cpp"
+#   "ir/*.cpp"
+#   "module_info.cpp"
+# )
