+++ src/tools/clair/codegen/fnt_tools.cpp
