+++ src/tools/clair/codegen/templates/module.txt
